build logYield functionality for Pair -> token

+++ contracts/rubiconPoolsv0/Strategy.sol
@@ -174,7 +174,7 @@ contract Strategy {
         uint256 askDenominator,
         uint256 bidNumerator,
         uint256 bidDenominator
-    ) external onlyPairs returns (uint, uint) {
+    ) external onlyPairs returns (uint256, uint256) {
         // main function to chain the actions of a single strategic market making transaction
         require(askNumerator > 0);
         require(askDenominator > 0);
